pyproject.toml
[project]
name = "nautobot-nvdatamodels"
version = "0.16.0"
description = "A Nautobot app that provides tools for modeling NVIDIA products and resources, such as NVLink domains and compute clusters."
authors = [
    {name = "Austin de Coup-Crank", email = "adecoupcrank@nvidia.com"},
    {name = "Spencer Amann", email = "samann@nvidia.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Intended Audience :: Information Technology",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "nautobot==2.4.20",
    "pydantic>=2.0.0",
]
keywords = ["nautobot", "nautobot-app", "nautobot-plugin", "nvidia"]

[project.optional-dependencies]
dev = [
    "coverage",
    "django-debug-toolbar==5.2.0",
    "invoke",
    "ipython",
    "pylint",
    "pylint-django",
    "pylint-nautobot",
    "ruff==0.5.5",
    "yamllint",
    "toml",
    "Markdown",
    "markdown-version-annotations==1.0.1",
    "mkdocs==1.6.0",
    "mkdocs-material==9.5.32",
    "mkdocstrings==0.25.2",
    "mkdocstrings-python==1.10.8",
    "griffe==1.1.1",
    "towncrier~=23.6.0",
    "to-json-schema",
    "jsonschema",
    "selenium",
    "splinter"
]

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins = "pylint_django, pylint_nautobot"
ignore = ".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx = "^(_|test_|Meta$)"

[tool.pylint.messages_control]
disable = """,
    line-too-long
"""

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.pylint-nautobot]
supported_nautobot_versions = ["2.4.20"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = [
    "D", # pydocstyle
    "F",
    "E",
    "W", # flake8
    "S", # bandit
    "I", # isort
]
ignore = [
    # warning: `one-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
    "D203", # 1 blank line required before class docstring

    # D212 is enabled by default in google convention, and complains if we have a docstring like:
    # """
    # My docstring is on the line after the opening quotes instead of on the same line as them.
    # """
    # We've discussed and concluded that we consider this to be a valid style choice.
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line

    # Produces a lot of issues in the current codebase.
    "D401", # First line of docstring should be in imperative mood
    "D407", # Missing dashed underline after section
    "D416", # Section name ends in colon
    "D107", # Missing docstring in `__init__`
    "E501", # Line too long
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"nautobot_nvdatamodels/migrations/*" = ["D"]
"nautobot_nvdatamodels/tests/*" = ["D", "S"]

[tool.djlint]
profile = "django"
ignore = "H021,T002"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nautobot_nvdatamodels"]

[tool.hatch.build.targets.sdist]
exclude = [
  ".uv-cache",
  ".venv",
  ".git",
  "*.pyc",
  "__pycache__",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "ipdb>=0.13.13",
    "python-dotenv>=1.1.0",
    "selenium>=4.32.0",
    "splinter>=0.21.0",
]

[tool.towncrier]
package = "nautobot_nvdatamodels"
directory = "changes"
filename = "docs/admin/release_notes/version_X.Y.md"
template = "development/towncrier_template.j2"
start_string = "<!-- towncrier release notes start -->"
issue_format = "[#{issue}](https://github.com/NVIDIA/nautobot-app-nvdatamodels/issues/{issue})"

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "dependencies"
name = "Dependencies"
showcontent = true

[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "housekeeping"
name = "Housekeeping"
showcontent = true
